server/requirements.txt
Django==3.2.11
PyMySQL==1.0.2
djangorestframework==3.13.0
django-cors-headers==3.13.0
Pillow==9.1.1
django-environ==0.10.0
stripe==7.8.2
requests==2.32.3